Automated System Design for Low Head River Diversion Structures
Hydrolics Software is Desktop application  Which is an automated
computer program for design of low head river diversion structure. The
software expected to help for hydraulic structures design for irrigation
system. And the software help designers automating the analysis of river
diversion structures typically with a broad crested type. It also
ensures the efficient distribution of the water supply at the right
place at the right amount and at the right time for irrigation land.
Authored requirements documentation, functional specifications, and
design documentation and programed code with C# language.